requirements.txt
cairocffi==1.2.0
CairoSVG==2.5.0
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
cssselect2==0.4.1
defusedxml==0.6.0
idna==2.10
Pillow==8.0.1
pony==0.7.14
psycopg2-binary==2.8.6
pycparser==2.20
requests==2.25.0
six==1.15.0
tinycss2==1.1.0
urllib3==1.26.2
vk-api==11.9.1
webencodings==0.5.1
